Fixed services and added defaults
Added commands to the defaults file and rebuild bindata.

Fixed a bug where it would always match filesystem as the service so it
couldn't play from any other sites.

Fixed displayed html in notification message so that if there is no URL
for the image then it doesn't try to display one.