Fixed playlists and listlocal
Did a bunch of work on the listlocal command so that it now works on
directories and playlists and outputs the correct information. Also
redid all the playlist stuff.

Added vendor stuff with the gvt command. I'm not sure if I did it
correctly and didn't include the manifest file. I'm not sure if I was
supposed to or not.

The version of github.com/jamesnetherton/m3u that was included here is a
version I modified and have in a fork JasonWaataja/m3u. I have made a
pull request but haven't gotten a response back yet.