Fixed two panics
The first was that there was a ++ instead of a -- which counted in the
wrong direction. Another was that there are portions in Skip that have
Queue[0] run with an empty queue, leading to an error because I believe
the way the program was made with concurrency with the reset command it
can clear the tracks and then attempt to skip with no tracks.