Started working but noticed problem
I assumed adding the createlocal feature would be easy until I realized
that it downloads in webm format for YouTube at least it seems. This
means that all the code I wrote for local flies won't work because it
was only designed to work with MP3s.

I'll figure out a solution to this later.